+++ modules/literal/src/main/scala/io/circe/validator/literal/JsontMacros.scala
@@ -1,0 +1,5 @@
+package io.circe.validator.literal
+
+import scala.reflect.macros.blackbox
+
+class JsontMacros(val c: blackbox.Context) extends ValidatorMacros
