+++ modules/scalatest-circe-validator/src/main/scala/org/scalatest/JsonValidatorMatchers.scala
@@ -1,13 +1,10 @@
 package org.scalatest
 
 import io.circe.Json
-import io.circe.validator.literal.ValidatorMacrosTrait
 import io.circe.validator.{Validator, run => runValidator}
 import org.scalatest.JsonValidatorPrettifier.jsonValidatorPrettifier
 import org.scalatest.matchers.{MatchResult, Matcher}
-
 import scala.language.experimental.macros
-import scala.reflect.macros.blackbox
 
 trait JsonValidatorMatchers {
 
@@ -31,17 +28,4 @@ trait JsonValidatorMatchers {
     def matchJson(args: Any*): Matcher[Json] =
       macro MatchJsonMacros.jsonMatcherStringContext
   }
-}
-
-class MatchJsonMacros(val c: blackbox.Context)
-    extends ValidatorMacrosTrait
-    with JsonValidatorMatchers {
-  import c.universe._
-
-  final def jsonMatcherStringContext(
-      args: c.Expr[Any]*
-  ): c.Expr[Matcher[Json]] =
-    c.Expr[Matcher[Json]](
-      q"passValidation(${validatorStringContextTree(args: _*)})"
-    )
 }
